Removed the autoloading helper and added a Mime::Type override
The helper was getting in the way since we have separate mobile and
application layouts.

The Mime::Type override sets the format earlier than a before callback.
This allows our FallbackResolver to work properly for all partials and
layouts (as rails was pulling format.request info prior to our before
callbacks).